internal: move PlatformPrefix to internal/linux
The platform prefix is tied to the Linux kernel, move it
to the appropriate package.
