+++ internal/linux/platform.go
@@ -1,4 +1,4 @@
-package internal
+package linux
 
 import (
 	"runtime"
